Implement subheader page title and update text component
